Explicitly cast double to float
The loss of precision is acceptable here. The cast removes the compiler warning.

+++ TreeGenerator/tree_generator_app.cpp
@@ -97,7 +97,7 @@ namespace tree_generator
 
 		void HandleScrollInput(CameraController* camera, double xOffset, double yOffset)
 		{
-			camera->GetCurrentMovement().remainingDistanceChange -= yOffset;
+			camera->GetCurrentMovement().remainingDistanceChange -= static_cast<float>(yOffset);
 		}
 
 		std::vector<lsystem::Symbol> CreateSimpleBinaryTree(
